requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
requests>=2.26.0,<3.0.0
aiohttp>=3.8.0,<4.0.0
python-dotenv>=0.19.0,<0.20.0
pyyaml>=6.0,<7.0
# For testing
pytest>=7.0.0,<8.0.0
